+++ test/nex/typechecker_test.clj
@@ -1197,15 +1197,23 @@ end"]
       (is (:success (tc/type-check (p/ast code)))))))
 
 (deftest test-string-to-bytes-types
-  (testing "String.to_bytes typechecks as Array[Integer]"
+  (testing "String.to_bytes typechecks as Array[Byte]"
+    (let [code "class Test
+  feature
+    demo() do
+      let xs: Array[Byte] := \"cat\".to_bytes()
+      let b: Byte := xs.get(1)
+    end
+end"]
+      (is (:success (tc/type-check (p/ast code))))))
+  (testing "String.to_bytes is no longer an Array[Integer]"
     (let [code "class Test
   feature
     demo() do
       let xs: Array[Integer] := \"cat\".to_bytes()
-      let b: Integer := xs.get(1)
     end
 end"]
-      (is (:success (tc/type-check (p/ast code)))))))
+      (is (not (:success (tc/type-check (p/ast code))))))))
 
 ;; Let type inference tests